Models/DB_Schemes/Data_Chunk.py
from pydantic import BaseModel ,Field, validator
from typing import Optional
from bson.objectid import ObjectId

class dataChunk():
    _id : Optional[ObjectId]
    chunk_text:str = Field(..., min_length=1)
    chunk_metadata :dict
    chunk_order : int = Field(..., gt=0)
    chunk_project_id : ObjectId



class config () :
    arbitrary_types_allowd = True
