pyproject.toml
[project]
name = "projec-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.14"
dependencies = [
    "aiofiles==23.2.1",
    "alembic==1.14.0",
    "asyncpg==0.30.0",
    "billiard==4.2.1",
    "celery==5.5.3",
    "cohere==5.5.8",
    "fastapi==0.110.2",
    "fastapi-health==0.4.0",
    "flower==2.0.1",
    "kombu==5.5.4",
    "langchain==0.1.20",
    "motor==3.4.0",
    "pymongo==4.8.0",
    "nltk==3.9.1",
    "openai==1.75.0",
    "pgvector==0.4.0",
    "prometheus-client==0.21.1",
    "psycopg2==2.9.10",
    "pydantic-mongo==2.3.0",
    "pydantic-settings==2.2.1",
    "pymupdf==1.24.3",
    "python-dotenv==1.0.1",
    "python-multipart==0.0.9",
    "qdrant-client==1.10.1",
    "redis==6.2.0",
    "sqlalchemy==2.0.36",
    "starlette-exporter==0.23.0",
    "uvicorn[standard]==0.29.0",
    "vine==5.1.0",
]
